% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtl.R
\name{vtlGetStructure}
\alias{vtlGetStructure}
\title{Show node structure}
\usage{
vtlGetStructure(sessionID, node)
}
\arguments{
\item{sessionID}{The symbolic name of an active VTL session}

\item{node}{The symbolic name of a node of the session}
}
\description{
Show identifiers, measures and attributes with data type of the input node.
}
\details{
This function returns the structure of a node.
}
\examples{
\dontrun{
  vtlAddStatements(session = 'test', statements = 'a := r_input;
                               b := 3;
                               c := abs(sqrt(14 + a));
                               d := a + b + c;
                               e := c + d / a;
                               f := e - d;
                               g := -f;
                               test := a * b + c / a;')
                               
  vtlGetStructure('test', 'test')
}
}
