% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtl.R
\name{vtlKillSessions}
\alias{vtlKillSessions}
\title{Kill sessions}
\usage{
vtlKillSessions(sessions)
}
\arguments{
\item{sessions}{or vector of names of the sessions to be removed}
}
\description{
Close and remove specified sessions from the VTL engine
}
\details{
This function is used to remove sessions from the engine.
}
\examples{
\dontrun{
  vtlAddStatements(session = 'test2', restartSession = T,
                  statements = 'a := r_input;
                               b := 3;
                               c := abs(sqrt(14 + a));
                               d := a + b + c;
                               e := c + d / a;
                               f := e - d;
                               g := -f;
                               test := a * b + c / a;')
 
  vtlListSessions()
  vtlKillSessions('test2')
  vtlListSessions()
}
}
