% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtl.R
\name{vtlListSessions}
\alias{vtlListSessions}
\title{List sessions}
\usage{
vtlListSessions()
}
\description{
List all sessions in the VTL engine
}
\details{
This function returns the list of sessions created in the engine.
}
\examples{
\dontrun{
  vtlAddStatements(session = 'test', restartSession = T,
                  statements = 'a := r_input;
                               b := 3;
                               c := abs(sqrt(14 + a));
                               d := a + b + c;
                               e := c + d / a;
                               f := e - d;
                               g := -f;
                               test := a * b + c / a;')
 
  vtlListSessions()
}
}
