% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtl.R
\name{vtlListStatements}
\alias{vtlListStatements}
\title{List VTL statements}
\usage{
vtlListStatements(sessionID)
}
\arguments{
\item{sessionID}{The symbolic name of an active VTL session}
}
\description{
Lists all statements defined in an existing VTL session.
}
\details{
This function returns a named list containing all the statements defined in the specified VTL session.
}
\examples{
\dontrun{
  vtlAddStatements(session = 'test', statements = 'a := r_input;
                               b := 3;
                               c := abs(sqrt(14 + a));
                               d := a + b + c;
                               e := c + d / a;
                               f := e - d;
                               g := -f;
                               test := a * b + c / a;')
                               
  vtlListStatements('test')
}
}
