% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtl.R
\name{vtlLogLevel}
\alias{vtlLogLevel}
\title{Sets the VTL backend logging level}
\usage{
vtlLogLevel(level)
}
\arguments{
\item{level}{the log level as one of the following strings: 'fatal' 'warn' 'info' 'debug' 'trace' 'off'}
}
\description{
This command sets a given logging level in the VTL Java engine.
}
