% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VTLSessionManager.R
\docType{data}
\name{VTLSessionManager}
\alias{VTLSessionManager}
\title{Manage VTL sessions}
\format{
An object of class \code{VTLSessionManager} (inherits from \code{R6}) of length 9.
}
\usage{
VTLSessionManager
}
\description{
R6 singleton object to find and destroy VTL sessions
}
\details{
This object is a preallocated instance of the class \code{\link{VTLSessionManagerClass}}.
See the class documentation page to help on the methods of this object.
}
\keyword{datasets}
