% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtl.R
\name{vtlStudio}
\alias{vtlStudio}
\title{Launch VTL editor}
\usage{
vtlStudio(...)
}
\arguments{
\item{...}{More options to pass to \code{\link[shiny]{runApp}}.}
}
\description{
This command opens a shiny-based rich editor for writing and testing VTL programs.
}
\examples{
\dontrun{
    #opens the editor 
    vtlStudio()
}
}
