% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtl.R
\name{vtlAddStatements}
\alias{vtlAddStatements}
\title{Process VTL statements}
\usage{
vtlAddStatements(sessionID, statements, restartSession = F)
}
\arguments{
\item{sessionID}{The symbolic name of an active VTL session}

\item{statements}{The code to be added to the session}

\item{restartSession}{\code{TRUE} if session must be restarted (default \code{FALSE})}
}
\description{
Replaces or adds more statements to an existing VTL session.
}
\details{
If you are replacing one or more already defined rules, 
         you need to set \code{restartSession} to \code{TRUE} to avoid errors.

         Always returns true.
}
\examples{
\dontrun{
  vtlAddStatements(session = 'test', restartSession = T,
                   statements = 'a := r_input;
                                  b := 3;
                                  c := abs(sqrt(14 + a));
                                  d := a + b + c;
                                  e := c + d / a;
                                  f := e - d;
                                  g := -f;
                                  test := a * b + c / a;')
  }
}
