% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VTLSession.R
\name{VTLSession}
\alias{VTLSession}
\title{Manage VTL sessions}
\description{
VTL Session
}
\details{
R6 Class for interacting with a VTL session instance.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of this VTL session.}

\item{\code{text}}{The temporary code buffer of this VTL session used by the editor.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{VTLSession$new()}}
\item \href{#method-finalize}{\code{VTLSession$finalize()}}
\item \href{#method-isCompiled}{\code{VTLSession$isCompiled()}}
\item \href{#method-print}{\code{VTLSession$print()}}
\item \href{#method-setText}{\code{VTLSession$setText()}}
\item \href{#method-addStatements}{\code{VTLSession$addStatements()}}
\item \href{#method-compile}{\code{VTLSession$compile()}}
\item \href{#method-getStatements}{\code{VTLSession$getStatements()}}
\item \href{#method-getNodes}{\code{VTLSession$getNodes()}}
\item \href{#method-getValues}{\code{VTLSession$getValues()}}
\item \href{#method-getTopology}{\code{VTLSession$getTopology()}}
\item \href{#method-clone}{\code{VTLSession$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new VTL session with a given name.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VTLSession$new(name = character(0))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name to identify this session}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
This method should not be called by the application.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
Terminates this VTL session.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VTLSession$finalize()}\if{html}{\out{</div>}}
}

\subsection{Details}{
This method should not be called by the application.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isCompiled"></a>}}
\if{latex}{\out{\hypertarget{method-isCompiled}{}}}
\subsection{Method \code{isCompiled()}}{
Check if this session was compiled.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VTLSession$isCompiled()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Returns \code{TRUE} if this VTL session has already been compiled.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Overrides the default print behaviour.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VTLSession$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setText"></a>}}
\if{latex}{\out{\hypertarget{method-setText}{}}}
\subsection{Method \code{setText()}}{
Changes the editor text in the session buffer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VTLSession$setText(code)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{code}}{The editor code to associate this session}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addStatements"></a>}}
\if{latex}{\out{\hypertarget{method-addStatements}{}}}
\subsection{Method \code{addStatements()}}{
Replace or add new VTL statements to this session and updates the session code buffer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VTLSession$addStatements(statements, restart = T)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{statements}}{The code to add to this session}

\item{\code{restart}}{TRUE if old code must be discarded before adding the new}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compile"></a>}}
\if{latex}{\out{\hypertarget{method-compile}{}}}
\subsection{Method \code{compile()}}{
Compiles the VTL statements submitted for this session.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VTLSession$compile()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getStatements"></a>}}
\if{latex}{\out{\hypertarget{method-getStatements}{}}}
\subsection{Method \code{getStatements()}}{
Obtains a named list of all the VTL statements submitted for this session.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VTLSession$getStatements()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getNodes"></a>}}
\if{latex}{\out{\hypertarget{method-getNodes}{}}}
\subsection{Method \code{getNodes()}}{
Obtains a named list of all rules and values submitted for this session.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VTLSession$getNodes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getValues"></a>}}
\if{latex}{\out{\hypertarget{method-getValues}{}}}
\subsection{Method \code{getValues()}}{
Returns a list of data frames containing the values of the named nodes defined in this session.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VTLSession$getValues(nodes)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nodes}}{a list of names of nodes to compute from this session}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getTopology"></a>}}
\if{latex}{\out{\hypertarget{method-getTopology}{}}}
\subsection{Method \code{getTopology()}}{
Creates a fore network representation of all nodes defined in this VTL session.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VTLSession$getTopology(distance = 100, charge = -100)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{distance}}{The distance between dots}

\item{\code{charge}}{The repelling force between dots}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VTLSession$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
