% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VTLSessionManager.R
\name{VTLSessionManagerClass}
\alias{VTLSessionManagerClass}
\title{Manage VTL sessions}
\description{
R6 class to create, find and destroy VTL sessions
}
\details{
This R6 class is used to create, find and destroy VTL sessions
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{VTLSessionManagerClass$new()}}
\item \href{#method-finalize}{\code{VTLSessionManagerClass$finalize()}}
\item \href{#method-find}{\code{VTLSessionManagerClass$find()}}
\item \href{#method-list}{\code{VTLSessionManagerClass$list()}}
\item \href{#method-clear}{\code{VTLSessionManagerClass$clear()}}
\item \href{#method-kill}{\code{VTLSessionManagerClass$kill()}}
\item \href{#method-getOrCreate}{\code{VTLSessionManagerClass$getOrCreate()}}
\item \href{#method-clone}{\code{VTLSessionManagerClass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new manager instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VTLSessionManagerClass$new()}\if{html}{\out{</div>}}
}

\subsection{Details}{
This method should not be called by the application.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
Finalization
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VTLSessionManagerClass$finalize()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Clears any managed VTL session when this manager is discarded by the garbage collector.
This method should not be called by the application.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-find"></a>}}
\if{latex}{\out{\hypertarget{method-find}{}}}
\subsection{Method \code{find()}}{
Returns the active session or an error if the search fails or a session doesn't exist.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VTLSessionManagerClass$find(sessionID)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sessionID}}{The name of the session to search}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list"></a>}}
\if{latex}{\out{\hypertarget{method-list}{}}}
\subsection{Method \code{list()}}{
List all active named VTL sessions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VTLSessionManagerClass$list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clear"></a>}}
\if{latex}{\out{\hypertarget{method-clear}{}}}
\subsection{Method \code{clear()}}{
All active VTL sessions are killed and a new VTL session named 'test' is created.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VTLSessionManagerClass$clear()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-kill"></a>}}
\if{latex}{\out{\hypertarget{method-kill}{}}}
\subsection{Method \code{kill()}}{
Silently terminates the named active VTL session if it exists.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VTLSessionManagerClass$kill(sessionID)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sessionID}}{The name of the session to kill}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getOrCreate"></a>}}
\if{latex}{\out{\hypertarget{method-getOrCreate}{}}}
\subsection{Method \code{getOrCreate()}}{
If the named VTL session exists, return it, otherwise create a new VTL session with the given name.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VTLSessionManagerClass$getOrCreate(sessionID)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sessionID}}{The name of the session to create}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VTLSessionManagerClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
